Add in both HlmsUnlit and HlmsPbs a static "getDefaultPaths" method to get the subfolder need to create the archive for each Hlms
Theses paths are used to create the actuall full path to the Hlms data directory when loading resources.
This permit the client code to not be broken if the organization of the files in the default Ogre distribution has been changed.
The GraphicsSystem class that is part of the sample/tutorial framework has been updated to uses them instead of manually setting all the paths.

--HG--
branch : v2-1